+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## v0.3.1 [unrelease]
+## v0.3.1 [2018-04-18]
 
 ### Release Notes
 
 - [#99](https://github.com/lodastack/agent/pull/99): only listen on `localhost` for trace
 - [#101](https://github.com/lodastack/agent/pull/101): support IPv6 address report
+- [#105](https://github.com/lodastack/agent/pull/105): add agent traffic API
 
 ### Bugfixes